requirements.txt
obs-websocket-py
simpleaudio
black
requests
pytchat
openai
opencv-python